Don't register CompletionStage's generic type for reflection if not necessary
